requirements.txt
python-telegram-bot==22.1
requests
numpy
scikit-learn
pillow
matplotlib
python-dotenv
